move check_vendor_specific_psustatus to test_platform_info.py
